#41: Split up mcp_server.py (#42)
