+++ doc/changes/unreleased.md
@@ -1,1 +1,5 @@
 # Unreleased
+
+## Refactoring
+
+* #41: Moved the `main` and `_register_tools` function to a separate file.
