+++ exasol/ai/mcp/server/main.py
@@ -1,0 +1,243 @@
+import json
+import os
+import re
+
+import pyexasol
+from pydantic import ValidationError
+from pyexasol import ExaConnection
+
+from exasol.ai.mcp.server.mcp_server import ExasolMCPServer
+from exasol.ai.mcp.server.server_settings import McpServerSettings
+
+ENV_SETTINGS = "EXA_MCP_SETTINGS"
+ENV_DSN = "EXA_DSN"
+ENV_USER = "EXA_USER"
+ENV_PASSWORD = "EXA_PASSWORD"
+
+
+def _register_list_schemas(mcp_server: ExasolMCPServer) -> None:
+    mcp_server.tool(
+        mcp_server.list_schemas,
+        description=(
+            "The tool lists schemas in the Exasol Database. "
+            "For each schema, it provides the name and an optional comment."
+        ),
+    )
+
+
+def _register_find_schemas(mcp_server: ExasolMCPServer) -> None:
+    mcp_server.tool(
+        mcp_server.find_schemas,
+        description=(
+            "The tool finds schemas in the Exasol Database by looking for the "
+            "specified keywords in their names and comments. "
+            "For each schema it finds, it provides the name and an optional "
+            "comment."
+        ),
+    )
+
+
+def _register_list_tables(mcp_server: ExasolMCPServer) -> None:
+    mcp_server.tool(
+        mcp_server.list_tables,
+        description=(
+            "The tool lists tables and views in the specified schema of the "
+            "the Exasol Database. For each table and view, it provides the "
+            "name and an optional comment."
+        ),
+    )
+
+
+def _register_find_tables(mcp_server: ExasolMCPServer) -> None:
+    mcp_server.tool(
+        mcp_server.find_tables,
+        description=(
+            "The tool finds tables and views in the Exasol Database by looking "
+            "for the specified keywords in their names and comments. "
+            "For each table or view it finds, it provides the name and an "
+            "optional comment. An optional `schema_name` argument allows "
+            "restricting the search to tables and views in the specified schema."
+        ),
+    )
+
+
+def _register_list_functions(mcp_server: ExasolMCPServer) -> None:
+    mcp_server.tool(
+        mcp_server.list_functions,
+        description=(
+            "The tool lists functions in the specified schema of the Exasol "
+            "Database. For each function, it provides the name and an optional "
+            "comment."
+        ),
+    )
+
+
+def _register_find_functions(mcp_server: ExasolMCPServer) -> None:
+    mcp_server.tool(
+        mcp_server.find_functions,
+        description=(
+            "The tool finds functions in the Exasol Database by looking for "
+            "the specified keywords in their names and comments. For each "
+            "function it finds, it provides the name and an optional comment. "
+            "An optional `schema_name` argument allows restricting the search "
+            "to functions in the specified schema."
+        ),
+    )
+
+
+def _register_list_scripts(mcp_server: ExasolMCPServer) -> None:
+    mcp_server.tool(
+        mcp_server.list_scripts,
+        description=(
+            "The tool lists the user defined functions (UDF) in the specified "
+            "schema of the Exasol Database. For each UDF, it provides the name "
+            "and an optional comment."
+        ),
+    )
+
+
+def _register_find_scripts(mcp_server: ExasolMCPServer) -> None:
+    mcp_server.tool(
+        mcp_server.find_scripts,
+        description=(
+            "The tool finds the user defined functions (UDF) in the Exasol "
+            "Database by looking for the specified keywords in their names and "
+            "comments. For each UDF it finds, it provides the name and an "
+            "optional comment. An optional `schema_name` argument allows "
+            "restricting the search to UDFs in the specified schema."
+        ),
+    )
+
+
+def _register_describe_table(mcp_server: ExasolMCPServer) -> None:
+    mcp_server.tool(
+        mcp_server.describe_table,
+        description=(
+            "The tool describes the specified table or view in the specified "
+            "schema of the Exasol Database. The description includes the list "
+            "of columns and for a table also the list of constraints. For each "
+            "column the tool provides the name, the SQL data type and an "
+            "optional comment. For each constraint it provides its type, e.g. "
+            "PRIMARY KEY, the list of columns the constraint is applied to and "
+            "an optional name. For a FOREIGN KEY it also provides the referenced "
+            "schema, table and a list of columns in the referenced table."
+        ),
+    )
+
+
+def _register_describe_function(mcp_server: ExasolMCPServer) -> None:
+    mcp_server.tool(
+        mcp_server.describe_function,
+        description=(
+            "The tool describes the specified function in the specified schema "
+            "of the Exasol Database. It provides the list of input parameters "
+            "and the return SQL type. For each parameter it specifies the name "
+            "and the SQL type."
+        ),
+    )
+
+
+def _register_describe_script(mcp_server: ExasolMCPServer) -> None:
+    mcp_server.tool(
+        mcp_server.describe_script,
+        description=(
+            "The tool describes the specified user defined function (UDF) in "
+            "the specified schema of the Exasol Database. It provides the "
+            "list of input parameters, the list of emitted parameters or the "
+            "SQL type of a single returned value. For each parameter it "
+            "provides the name and the SQL type. Both the input and the "
+            "emitted parameters can be dynamic or, in other words, flexible. "
+            "The dynamic parameters are indicated with ... (triple dot) string "
+            "instead of the parameter list. The description includes some usage "
+            "notes and a call example."
+        ),
+    )
+
+
+def _register_execute_query(mcp_server: ExasolMCPServer) -> None:
+    mcp_server.tool(
+        mcp_server.execute_query,
+        description=(
+            "The tool executes the specified query in the Exasol Database. The "
+            "query must be a SELECT statement. The tool returns data selected "
+            "by the query."
+        ),
+    )
+
+
+def register_tools(mcp_server: ExasolMCPServer, config: McpServerSettings) -> None:
+    if config.schemas.enable:
+        _register_list_schemas(mcp_server)
+        _register_find_schemas(mcp_server)
+    if config.tables.enable or config.views.enable:
+        _register_list_tables(mcp_server)
+        _register_find_tables(mcp_server)
+    if config.functions.enable:
+        _register_list_functions(mcp_server)
+        _register_find_functions(mcp_server)
+    if config.scripts.enable:
+        _register_list_scripts(mcp_server)
+        _register_find_scripts(mcp_server)
+    if config.columns.enable:
+        _register_describe_table(mcp_server)
+    if config.parameters.enable:
+        _register_describe_function(mcp_server)
+        _register_describe_script(mcp_server)
+    if config.enable_read_query:
+        _register_execute_query(mcp_server)
+
+
+def get_mcp_settings() -> McpServerSettings:
+    """
+    Reads optional settings. They can be provided either in a json string stored in the
+    EXA_MCP_SETTINGS environment variable or in a json file. In the latter case
+    EXA_MCP_SETTINGS must contain the file path.
+    """
+    try:
+        settings_text = os.environ.get(ENV_SETTINGS)
+        if not settings_text:
+            return McpServerSettings()
+        elif re.match(r"^\s*\{.*\}\s*$", settings_text):
+            return McpServerSettings.model_validate_json(settings_text)
+        elif os.path.isfile(settings_text):
+            with open(settings_text) as f:
+                return McpServerSettings.model_validate(json.load(f))
+        raise ValueError(
+            "Invalid MCP Server configuration settings. The configuration "
+            "environment variable should either contain a json string or "
+            "point to an existing json file."
+        )
+    except (ValidationError, json.decoder.JSONDecodeError) as config_error:
+        raise ValueError("Invalid MCP Server configuration settings.") from config_error
+
+
+def create_mcp_server(
+    connection: ExaConnection, config: McpServerSettings
+) -> ExasolMCPServer:
+    """
+    Creates the Exasol MCP Server and registers its tools.
+    """
+    mcp_server = ExasolMCPServer(connection=connection, config=config)
+    register_tools(mcp_server, config)
+    return mcp_server
+
+
+def main():
+    """
+    Main entry point that creates and runs the MCP server.
+    """
+    dsn = os.environ[ENV_DSN]
+    user = os.environ[ENV_USER]
+    password = os.environ[ENV_PASSWORD]
+    mcp_settings = get_mcp_settings()
+
+    connection = pyexasol.connect(
+        dsn=dsn, user=user, password=password, fetch_dict=True, compression=True
+    )
+
+    mcp_server = create_mcp_server(connection, mcp_settings)
+    mcp_server.run()
+
+
+if __name__ == "__main__":
+    main()
