+++ pyproject.toml
@@ -12,7 +12,7 @@ keywords = ['exasol', 'MCP server']
 dynamic = ["dependencies"]
 
 [project.scripts]
-exasol-mcp-server = "exasol.ai.mcp.server.mcp_server:main"
+exasol-mcp-server = "exasol.ai.mcp.server.main:main"
 
 [tool.poetry]
 requires-poetry = ">=2.1.0"
