+++ test/integration/test_mcp_server.py
@@ -19,10 +19,8 @@ from fastmcp import Client
 from fastmcp.exceptions import ToolError
 from pyexasol import ExaConnection
 
-from exasol.ai.mcp.server.mcp_server import (
-    TABLE_USAGE,
-    ExasolMCPServer,
-)
+from exasol.ai.mcp.server.main import create_mcp_server
+from exasol.ai.mcp.server.mcp_server import TABLE_USAGE
 from exasol.ai.mcp.server.parameter_parser import FUNCTION_USAGE
 from exasol.ai.mcp.server.server_settings import (
     ExaDbResult,
@@ -36,13 +34,13 @@ from exasol.ai.mcp.server.server_settings import (
 async def _run_tool_async(
     connection: ExaConnection, config: McpServerSettings, tool_name: str, **kwargs
 ):
-    exa_server = ExasolMCPServer(connection, config)
+    exa_server = create_mcp_server(connection, config)
     async with Client(exa_server) as client:
         return await client.call_tool(tool_name, kwargs)
 
 
 async def _list_tools_async(connection: ExaConnection, config: McpServerSettings):
-    exa_server = ExasolMCPServer(connection, config)
+    exa_server = create_mcp_server(connection, config)
     async with Client(exa_server) as client:
         return await client.list_tools()
